Rewritten with es6, fixed some typos, added pre-publish step
